+++ include/boost/simd/meta/abi_of.hpp
@@ -11,8 +11,8 @@
 
 **/
 //==================================================================================================
-#ifndef BOOST_SIMD_SDK_ABI_OF_HPP_INCLUDED
-#define BOOST_SIMD_SDK_ABI_OF_HPP_INCLUDED
+#ifndef BOOST_SIMD_META_ABI_OF_HPP_INCLUDED
+#define BOOST_SIMD_META_ABI_OF_HPP_INCLUDED
 
 #include <boost/simd/config.hpp>
 #include <boost/simd/arch/spec.hpp>
